*** empty log message ***

+++ trunk/tests/test_config.h
@@ -144,15 +144,15 @@ typedef size_t KEY;
   ACE_OS::unlink (temp);
 #endif /* ! VXWORKS */
 
-const int ACE_NS_MAX_ENTRIES = 1000;
-const int ACE_DEFAULT_USECS = 1000;
-const int ACE_MAX_TIMERS = 4;
-const int ACE_MAX_THREADS = 4;
-const int ACE_MAX_DELAY = 10;
-const int ACE_MAX_INTERVAL = 0;
-const int ACE_MAX_ITERATIONS = 10;
-const int ACE_MAX_PROCESSES = 10;
-const int ACE_MAX_CLIENTS = 30;
+const size_t ACE_NS_MAX_ENTRIES = 1000;
+const size_t ACE_DEFAULT_USECS = 1000;
+const size_t ACE_MAX_TIMERS = 4;
+const size_t ACE_MAX_THREADS = 4;
+const size_t ACE_MAX_DELAY = 10;
+const size_t ACE_MAX_INTERVAL = 0;
+const size_t ACE_MAX_ITERATIONS = 10;
+const size_t ACE_MAX_PROCESSES = 10;
+const size_t ACE_MAX_CLIENTS = 30;
 
 char ACE_ALPHABET[] = "abcdefghijklmnopqrstuvwxyz";
